# server/docs/swagger.yaml
basePath: /ttms
definitions:
  common.State:
    properties:
      data:
        description: 失败时返回空
      status_code:
        description: 状态码，0-成功，其他值-失败
        type: integer
      status_msg:
        description: 返回状态描述
        type: string
    type: object
...
  reply.GetTags:
    properties:
      tags_names:
        items:
          type: string
        type: array
    type: object
  reply.GetTagsInMovie:
    properties:
      tag_names:
        items:
          type: string
        type: array
    type: object
...
  request.DeleteOneByMovieAndTag:
    properties:
      MovieId:
        minimum: 1
        type: integer
      TagNames:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - MovieId
    - TagNames
    type: object
...
host: ttms.humraja.xyz
info:
  contact: {}
  description: ttms影院管理系统
  license:
    name: raja,moonman
  title: ttms
  version: "1.0"
paths:
...
  /tags/addTags:
    post:
      consumes:
      - application/json
      parameters:
      - description: 电影Id，要添加的标签名数组
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteOneByMovieAndTag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 为电影添加标签
      tags:
      - tags
  /tags/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 电影Id，标签名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteOneByMovieAndTag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 删除电影的指定标签
      tags:
      - tags
  /tags/list:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        minimum: 1
        name: MovieId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetTagsInMovie'
              type: object
      security:
      - BasicAuth: []
      summary: 得到电影相关的所有标签
      tags:
      - tags
  /tags/list/tags:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetTags'
              type: object
      security:
      - BasicAuth: []
      summary: 得到所有标签
      tags:
      - tags
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"